fixed how ItemsAdder integration is handled to avoid error if the plugin is not installed on the server
