+++ eco-core/core-plugin/src/main/java/com/willfp/ecoskills/EcoSkillsPlugin.java
@@ -16,6 +16,7 @@ import com.willfp.ecoskills.data.LeaderboardHandler;
 import com.willfp.ecoskills.effects.Effect;
 import com.willfp.ecoskills.effects.Effects;
 import com.willfp.ecoskills.integrations.EcoEnchantsEnchantingLeveller;
+import com.willfp.ecoskills.integrations.ItemsAdderMiningLeveller;
 import com.willfp.ecoskills.skills.Skill;
 import com.willfp.ecoskills.skills.SkillDisplayListener;
 import com.willfp.ecoskills.skills.SkillLevellingListener;
@@ -125,7 +126,8 @@ public class EcoSkillsPlugin extends EcoPlugin {
     @Override
     protected List<IntegrationLoader> loadIntegrationLoaders() {
         return Arrays.asList(
-                new IntegrationLoader("EcoEnchants", () -> this.getEventManager().registerListener(new EcoEnchantsEnchantingLeveller(this)))
+                new IntegrationLoader("EcoEnchants", () -> this.getEventManager().registerListener(new EcoEnchantsEnchantingLeveller(this))),
+                new IntegrationLoader("ItemsAdder", () -> this.getEventManager().registerListener(new ItemsAdderMiningLeveller(this)))
         );
     }
