+++ eco-core/core-plugin/src/main/kotlin/com/willfp/ecoskills/skills/skills/SkillMining.kt
@@ -16,11 +16,14 @@ class SkillMining : Skill(
     "mining"
 ) {
     private val rewards: MutableMap<Material, Double>
-    private val iarewards: MutableMap<String, Double>
+
+    companion object {
+        val iarewards: MutableMap<String, Double> = mutableMapOf()
+    }
+
 
     init {
         rewards = EnumMap(Material::class.java)
-        iarewards = mutableMapOf()
     }
 
     override fun postUpdate() {
@@ -50,28 +53,6 @@ class SkillMining : Skill(
         val toGive = rewards[type] ?: return
 
         if (BlockUtils.isPlayerPlaced(event.block)) {
-            return
-        }
-
-        if (plugin.configYml.getBool("skills.prevent-levelling-while-afk") && AFKManager.isAfk(player)) {
-            return
-        }
-
-        player.giveSkillExperience(this, toGive)
-    }
-
-    @EventHandler(ignoreCancelled = true, priority = EventPriority.MONITOR)
-    fun handleLevelling(event: CustomBlockBreakEvent){
-        val type = event.namespacedID
-        val player = event.player
-
-        if(player.gameMode == GameMode.CREATIVE || player.gameMode == GameMode.SPECTATOR){
-            return
-        }
-
-        val toGive = iarewards[type] ?: return
-
-        if(BlockUtils.isPlayerPlaced(event.block)){
             return
         }
